Logger: use better colors on Linux
Terminal backgrounds may not be always black, there are people
which prefer bright/white terminal backgrounds. Yellow and White
obiously aren't readable on such a background color.

The colors which are now used are universal to bright and dark
backgrounds and readable. For this the terminal emulator needs
to support 16-million (24-bit) colors. If that mode is not
supported, the underlaying shell fallbacks to the default 8/16
colors by default.

As an extend the color is also reset to default on application
cleanup to prevent the last used color to affect the users shell.